Use memfd_create instead of posix shm for linux
For linux OSes use memfd_create instead of posix shm. Add a
force-posix-shm feature to force the posix shm implementation
to be used.

Note: glibc does not currently provide a wrapper for this call yet. As a
result, it is recommended to use `syscall` directly. See the man pages
for details.